Add existsById method and improve error handling in update and delete methods

+++ src/main/java/com/imb2025/calificaciones/service/jpa/AsignacionDocenteServiceImp.java
@@ -23,16 +23,27 @@ public class AsignacionDocenteServiceImp implements IAsignacionDocenteService {
         return repository.findById(id).orElse(null);
     }
 
+    public AsignacionDocente existsById(Long id) {
+        return repository.existsById(id) ? findById(id) : null;
+    }
+
     public AsignacionDocente save(AsignacionDocente asignacionDocente) {
         return repository.save(asignacionDocente);
     }
 
     public AsignacionDocente update(Long id, AsignacionDocente asignacionDocente) {
+        repository.findById(id).orElseThrow(
+                () -> new RuntimeException(
+                        "Can't update AsignacionDocente with id: " + id + " because it does not exist"));
         asignacionDocente.setId(id);
         return repository.save(asignacionDocente);
     }
 
     public void deleteById(Long id) {
+        if (!repository.existsById(id)) {
+            throw new RuntimeException(
+                    "Can't delete AsignacionDocente with id: " + id + " because it does not exist");
+        }
         repository.deleteById(id);
     }
 }
